refactor(layout): localize validation rules
Move input, process, and collaboration validation into their lifecycle owners and remove the generic validation seam.
