+++ lib/layout/collaboration/steps/validateCollaboration.js
@@ -1,4 +1,4 @@
-import { validateMessageFlows } from '../../bpmn/Validation.js';
+import { LayoutError } from '../../../LayoutError.js';
 
 /**
  * @typedef {import('../../Types.js').CollaborationLayoutContext} CollaborationLayoutContext
@@ -9,7 +9,15 @@ import { validateMessageFlows } from '../../bpmn/Validation.js';
  * @returns {CollaborationLayoutContext}
  */
 export function validateCollaboration(context) {
-  validateMessageFlows(context.collaboration.messageFlows || []);
+  for (const flow of context.collaboration.messageFlows || []) {
+    if (!flow.sourceRef || !flow.targetRef) {
+      throw new LayoutError(
+        'INVALID_MESSAGE_FLOW_ENDPOINT',
+        flow.id,
+        'A message flow must reference source and target interaction nodes.'
+      );
+    }
+  }
 
   return context;
 }
