+++ lib/layout/index.js
@@ -1,6 +1,6 @@
 import { BpmnModdle } from 'bpmn-moddle';
 
-import { is } from '../di/DiUtil.js';
+import { getDefaultSize, is } from '../di/DiUtil.js';
 import { LayoutError } from '../LayoutError.js';
 import { LayoutWarning } from '../LayoutWarning.js';
 import { generateDiagrams } from './DiagramGeneration.js';
@@ -9,10 +9,6 @@ import {
   isSupportedVisualElement,
   getExpandedIds
 } from './bpmn/Predicates.js';
-import {
-  validateInputVisuals,
-  validateParseWarnings
-} from './bpmn/Validation.js';
 import { layoutCollaboration } from './collaboration/index.js';
 import { layoutProcessScope } from './process/index.js';
 
@@ -135,4 +131,60 @@ class LayoutEngine {
 
 export function layoutProcess(xml) {
   return new LayoutEngine().layoutProcess(xml);
+}
+
+function validateInputVisuals(definitions, root) {
+  const diagram = (definitions.diagrams || []).find(candidate => {
+    return candidate.plane?.bpmnElement === root;
+  });
+
+  for (const di of diagram?.plane?.planeElement || []) {
+    if (!di.$instanceOf('bpmndi:BPMNShape')) {
+      continue;
+    }
+
+    const element = di.bpmnElement;
+
+    if (
+      !element ||
+      is(element, 'bpmn:Group') ||
+      (getDefaultSize(element) && isSupportedVisualElement(element))
+    ) {
+      continue;
+    }
+
+    throw new LayoutError(
+      'UNSUPPORTED_ELEMENT',
+      element.id,
+      `Cannot generate DI for visual BPMN element "${element.$type}".`
+    );
+  }
+}
+
+function validateParseWarnings(warnings, xml) {
+  const invalidProcessRef = warnings.find(warning => {
+    return warning.property === 'bpmn:processRef';
+  });
+
+  if (invalidProcessRef) {
+    throw new LayoutError(
+      'INVALID_PARTICIPANT_PROCESS_REFERENCE',
+      invalidProcessRef.element.id,
+      'A participant processRef must reference a declared process.'
+    );
+  }
+
+  const unknownType = warnings.find(warning => {
+    return warning.message.includes('unknown type');
+  });
+
+  if (unknownType) {
+    const elementId = /<bpmn:[^ >]+[^>]*\sid="([^"]+)"/.exec(xml)?.[1] || 'unknown';
+
+    throw new LayoutError(
+      'UNSUPPORTED_ELEMENT',
+      elementId,
+      'Cannot generate DI for an unknown BPMN visual element.'
+    );
+  }
 }
